java/src/main/java/com/yourtion/leetcode/daily/m11/d23/Solution.java
package com.yourtion.leetcode.daily.m11.d23;

import java.util.Arrays;
import java.util.Comparator;

/**
 * 452. 用最少数量的箭引爆气球
 *
 * @author Yourtion
 * @link https://leetcode-cn.com/problems/minimum-number-of-arrows-to-burst-balloons/
 */
public class Solution {
    public int findMinArrowShots(int[][] points) {
        if (points.length == 0) {
            return 0;
        }
        Arrays.sort(points, Comparator.comparingInt(point -> point[1]));
        int pos = points[0][1];
        int ans = 1;
        for (int[] balloon : points) {
            if (balloon[0] > pos) {
                pos = balloon[1];
                ++ans;
            }
        }
        return ans;
    }
}
